requirements.txt
requests
PyYAML
apscheduler
python-dotenv
# Google Cloud Libraries
google-generativeai
google-cloud-aiplatform
# E-commerce (for future features)
ShopifyAPI
